Create README.md
This is a security surveillance system for a household that will monitor all external activities and in the process be able to detect and capture facial and body features. When that occurs, the camera will start recording for a specific amount of time, even in the absence of the feature until time specified for recording to stop.